Merge CppNotes with profile page
I had a separate repo to keep notes about C++ and such. I had merged a
bunch of my other repos into this one, so why not include that one?
Seemed like a good idea. Not sure why I want to consolidate my repos,
but it's nice to have it all in one place! I would reference the other
repo in this commit, but I'm probably going to delete it.