Calculate vector length
I wasn't sure where to go with adding functionality to my linear algebra
library, so I figured I would add the dot product and use that to
calculate the vector's length.
I decided to make lynel a compiled library since vec3d isn't a template
and the functions can be compiled. Although, I will probably make it a
template in the future. (In fact, that may be my next step...)