Matrix is the root of linear algebra types
A vector is a matrix and I want them represented as such. I wasn't sure
how I felt about subclassing and whatnot and I have interacted with the
Eigen library where the vector class is really just an alias for the
matrix class with a single row. So, I tried that with this. It will make
for some rework and I know there will be problems since the
specializations are broken into their own files. I have some ideas for
that and this is a WIP.