Single vector specialization per extent
It bothered me that my previous implementation used two different
template specializations for the row vs column vectors. I didn't want to
take the same approach as linalg.h, because that doesn't differentiate
between the two. However, it was a pain to figure out SFINAE for class
template specialization. Not sure why I was having the trouble, but I
finally figured it out.